Remove `utilities` and move the code to already existing `util.py`
The rationale is that `alpha` already depends on `util.py`
because of `NeptuneException` so there is no need
to introduce more dependencies.

We will need to cleanup modules anyway.

# Conflicts:
#	src/neptune_fetcher/alpha/api_client.py
