+++ src/neptune_fetcher/api/api_client.py
@@ -81,8 +81,8 @@ from neptune_fetcher.fields import (
     FieldType,
     FloatPointValue,
 )
-from neptune_fetcher.util import NeptuneException
-from neptune_fetcher.utilities.net import (
+from neptune_fetcher.util import (
+    NeptuneException,
     create_auth_api_client,
     get_config_and_token_urls,
 )
